requirements.txt
pandas
numpy>=1.24.3
scikit-learn
python-dotenv==1.0.1
Flask==3.0.0
Flask-Cors==4.0.0
SQLAlchemy
psycopg[binary]
gunicorn==21.2.0
pydantic==2.8.2
requests==2.31.0
mlflow==2.14.1